@@ -1,18 +1,19 @@
 name: example
 description: A new Flutter project.
 version: 1.0.0+1
+publish_to: "none"
 
 environment:
-  sdk: ">=2.0.0-dev.68.0 <3.0.0"
+  sdk: ">=2.19.1 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
+  # flutter_sequence_animation: "^3.0.0"
   cupertino_icons: ^0.1.2
-  sweetalert:
+  flutter_sweet_alert:
     path: ../
-  flutter_sequence_animation: "^3.0.0"
 
 dev_dependencies:
   flutter_test:
