@@ -1,16 +1,15 @@
-name: sweetalert
+name: flutter_sweet_alert
 description: sweet alert for flutter.
 version: 0.0.1
-author: JZoom <jzoom8112@gmail.com>
-homepage: https://github.com/jzoom/sweetalert
+homepage: https://github.com/OscarTinajero117/flutter_sweet_alert
 
 environment:
-  sdk: ">=2.0.0-dev.68.0 <3.0.0"
+  sdk: ">=2.19.1 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  flutter_sequence_animation: any
+  flutter_sequence_animation: ^4.0.0
 
 dev_dependencies:
   flutter_test:
