@@ -1,6 +1,6 @@
 import 'package:flutter_test/flutter_test.dart';
 
-import 'package:sweetalert/sweetalert.dart';
+// import 'package:flutter_sweet_alert/flutter_sweet_alert.dart';
 
 void main() {
   test('adds one to input values', () {});
